Fix conversion from mbcs, avoid infinite loop (PR#17569).
git-svn-id: https://svn.r-project.org/R/trunk@77638 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/engine.c
@@ -1,6 +1,6 @@
 /*
  *  R : A Computer Language for Statistical Data Analysis
- *  Copyright (C) 2001-2018  The R Core Team.
+ *  Copyright (C) 2001-2020  The R Core Team.
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -1797,7 +1797,7 @@ void GEText(double x, double y, const char * const str, cetype_t enc,
 					wchar_t wc;
 					mbstate_t mb_st;
 					mbs_init(&mb_st);
-					while ((used = mbrtowc(&wc, ss, n, &mb_st)) > 0) {
+					while ((int)(used = mbrtowc(&wc, ss, n, &mb_st)) > 0) {
 #ifdef DEBUG_MI
 					    printf(" centring %s aka %d in MBCS\n", ss, wc);
 #endif
@@ -2627,7 +2627,7 @@ void GEStrMetric(const char *str, cetype_t enc, const pGEcontext gc,
                     wchar_t wc;
                     mbstate_t mb_st;
                     mbs_init(&mb_st);
-                    while ((used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
+                    while ((int)(used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
                         GEMetricInfo((int) wc, gc, &asc, &dsc, &wid, dd);
                         if (asc > *ascent)
                             *ascent = asc;
@@ -2688,7 +2688,7 @@ void GEStrMetric(const char *str, cetype_t enc, const pGEcontext gc,
                     wchar_t wc;
                     mbstate_t mb_st;
                     mbs_init(&mb_st);
-                    while ((used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
+                    while ((int)(used = mbrtowc(&wc, s, n, &mb_st)) > 0) {
                         GEMetricInfo((int) wc, gc, &asc, &dsc, &wid, dd);
                         if (dsc > *descent)
                             *descent = dsc;
